refactored query building and exporting logic to a separate Query abstraction
